+++ src/include/ssystem.h
@@ -90,11 +90,11 @@ private:
 
 	// helper function for sending current state to Python side
 	void dumpCurrentStateToPython(void);
-	void sendTrajectory_CurrentStateToPython(double current_time, int arrType = -77);
+	void sendTrajectory_CurrentStateToPython(double current_time, double arrType = -77.0);
 	void sendTransitionStateVectorToPython(boolvector transition_states, double current_time);
 
 	void exportTime(double& simTime, double& lastExportTime);
-	void exportInterval(double simTime, int period, int arrType = -88);
+	void exportInterval(double simTime, int period, double arrType = -88.0);
 	void exportTrajState(double simTime, double* lastExportTime, int period);
 
 	void printAllMoves(void);
